[WS-182] Fixes NPE for TempWarps * Fixes not auto creating language files on BungeeCord
This issue occurs while creating the name of a TempWarp. But this only a few times. If somebody toggles his own TempWarp from private to public and the name of this TempWarp is already taken then it will be set to null. If you create a new TempWarp in this moment (between toggling it to public and setting a new name) my plugin tries to search for TempWarps (to prevent doubled names) and got that name from above (null) and directly throws a NullPointerException.